+++ lua/obsidian/workspace.lua
@@ -1,3 +1,4 @@
+local Path = require "plenary.path"
 local abc = require "obsidian.abc"
 
 ---@class obsidian.Workspace : obsidian.ABC
@@ -12,38 +13,59 @@ local Workspace = abc.new_class {
 ---Create a new workspace
 ---
 ---@param name string Workspace name
----@param path string Workspace path (will be normalized)
----
+---@param path string|Path Workspace path (will be normalized)
 ---@return obsidian.Workspace
 Workspace.new = function(name, path)
   local self = Workspace.init()
   self.name = name
-  self.path = vim.fs.normalize(path)
+  self.path = vim.fs.normalize(tostring(path))
   return self
 end
 
+---@return obsidian.Workspace
 Workspace.new_from_cwd = function()
-  return Workspace.new_from_dir(vim.fn.getcwd())
-end
+  -- First traverse upwards to find the root of the obsidian vault.
+  -- If found, use that, otherwise use the current directory as is.
+  local vault_indicator_folder = ".obsidian"
+  local cwd = assert(vim.fn.getcwd())
+  local dirs = Path:new(cwd):parents()
+  table.insert(dirs, 1, cwd)
 
-Workspace.new_from_dir = function(dir)
-  return Workspace.new(vim.fn.fnamemodify(dir, ":t"), dir)
-end
-
----Determines if cwd is a workspace
----
----@param workspaces table<obsidian.Workspace>
----@return obsidian.Workspace|nil
-Workspace.get_workspace_from_cwd = function(workspaces)
-  local cwd = vim.fn.getcwd()
-  local _, value = next(vim.tbl_filter(function(w)
-    if w.path == cwd then
-      return true
+  for _, dir in ipairs(dirs) do
+    local maybe_vault = Path:new(dir) / vault_indicator_folder
+    if maybe_vault:is_dir() then
+      return Workspace.new_from_dir(dir)
     end
-    return false
-  end, workspaces))
+  end
 
-  return value
+  return Workspace.new_from_dir(cwd)
+end
+
+---@param dir string|Path
+---@return obsidian.Workspace
+Workspace.new_from_dir = function(dir)
+  return Workspace.new(assert(vim.fs.basename(tostring(dir))), dir)
+end
+
+---Get the workspace corresponding to the current working directory (or a parent of), if there
+---is one.
+---
+---@param workspaces obsidian.Workspace[]
+---@return obsidian.Workspace|?
+Workspace.get_workspace_from_cwd = function(workspaces)
+  local cwd = assert(vim.fn.getcwd())
+  local dirs = Path:new(cwd):parents()
+  table.insert(dirs, 1, cwd)
+
+  for _, w in ipairs(workspaces) do
+    for _, dir in ipairs(dirs) do
+      if w.path == dir then
+        return w
+      end
+    end
+  end
+
+  return nil
 end
 
 ---Returns the default workspace