+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added Lua API methods `Client:set_workspace()` and `Client:switch_workspace`.
+
+### Fixed
+
+- Made workspace API more robust.
+
 ## [v2.4.0](https://github.com/epwalsh/obsidian.nvim/releases/tag/v2.4.0) - 2023-12-19
 
 ### Added